Move 'batch_fastqs' from icell8_stats.py to icell8.utils module.
